Refactor: move DBService to a new db_service module and update imports across files
