+++ src/wembed/cli/__init__.py
@@ -1,5 +1,5 @@
 from wembed.config import AppConfig
-from wembed.db import DBService
+from wembed.services.db_service import DBService
 
 cli_config = AppConfig()
 cli_db_service = DBService(cli_config)
