+++ src/wembed/services/__init__.py
@@ -1,0 +1,1 @@
+from .db_service import DBService
